+++ examples/ethernet_network.rb
@@ -19,7 +19,7 @@ my_client = {
 oneview_ethernet_network 'Eth1' do
   client my_client
   data(
-    vlanId: '1001',
+    vlanId: 1001,
     purpose: 'General',
     smartLink: false,
     privateNetwork: false
@@ -31,7 +31,7 @@ end
 oneview_ethernet_network 'Eth1' do
   client my_client
   data(
-    vlanId: '1001',
+    vlanId: 1001,
     purpose: 'General',
     smartLink: false,
     privateNetwork: false