+++ libraries/resource_providers/api300/c7000/ethernet_network_provider.rb
@@ -1,7 +1,3 @@
-#
-# Cookbook Name:: oneview_test
-# Attributes:: default
-#
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,9 +8,15 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
-#
 
-default['oneview']['api_version'] = 200
-# default['oneview']['api_variant'] = 'C7000'
+require_relative '../../api200/ethernet_network_provider'
 
-default['oneview_test']['client'] = { url: 'https://oneview.example.com', user: 'Administrator', password: 'secret123', api_version: 200 }
+module OneviewCookbook
+  module API300
+    module C7000
+      # EthernetNetworkProvider
+      class EthernetNetworkProvider < OneviewCookbook::API200::EthernetNetworkProvider
+      end
+    end
+  end
+end