+++ spec/unit/libraries/oneview_resource_base_spec.rb
@@ -7,7 +7,14 @@ RSpec.describe OneviewCookbook::ResourceBaseProperties do
     before :each do
       @context = {}
       allow(@context).to receive(:property).and_return true
-      allow(@context).to receive(:node).and_return('oneview' => { 'save_resource_info' => true })
+      attributes = {
+        'oneview' => {
+          'save_resource_info' => true,
+          'api_version' => 200,
+          'api_variant' => 'C7000'
+        }
+      }
+      allow(@context).to receive(:node).and_return(attributes)
     end
 
     it 'creates the :client property' do
@@ -27,6 +34,22 @@ RSpec.describe OneviewCookbook::ResourceBaseProperties do
 
     it 'creates the :save_resource_info property' do
       expect(@context).to receive(:property).with(:save_resource_info, [TrueClass, FalseClass, Array], default: true)
+      described_class.load(@context)
+    end
+
+    it 'creates the :api_version property' do
+      expect(@context).to receive(:property).with(:api_version, Fixnum, default: @context.node['oneview']['api_version'])
+      described_class.load(@context)
+    end
+
+    it 'creates the :api_variant property' do
+      expect(@context).to receive(:property)
+        .with(:api_variant, [String, Symbol], default: @context.node['oneview']['api_variant'])
+      described_class.load(@context)
+    end
+
+    it 'creates the :api_header_version property' do
+      expect(@context).to receive(:property).with(:api_header_version, Fixnum)
       described_class.load(@context)
     end
   end