+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+  - Support oneview-sdk v3.0 & different API versions/modules
+  - ethernet_network support for API300
+
 ### 1.1.0
   - Add support for client ENV variables
   - Fixed volume resource (#92) & examples